Final Objective: AI Malware Injection
Adds a new final objective, wherein the traitor must hack APCs with a malware injection device, and then build a special computer in a specific department (medical, science, engineering, or cargo). Hacking an APC with the device takes 20 seconds, and hacked APCs look identical to APCs that have been hacked by an emag or a malf AI. Using the injector on the computer in the specified department will set off every fire alarm in the department, close and bolt the doors to the room the computer is in, and begin a process wherein the AI slowly becomes corrupted over the course of two minutes, turning its current laws into unreadable gibberish. This process is very obvious, as the AI will automatically announce its corruption progress over the radio with every stage. Once the AI is 100% corrupted, it will turn into a malfunctioning AI with the sole objective of destroying the station with a nuclear device (although, like all objectives, this is more of a suggestion than a rule). It will not be loyal to anyone except itself. The AI will receive 10 processing power for every APC that the traitor hacked before uploading the malware. The minimum requirement for APCs is 5, but it's strongly recommended that traitors hack more for the extra processing power, because a malf AI takes 60 seconds to hack APCs and can't do anything during that time.

This objective can only appear if there is already a living, active AI in the game. As always, the numbers might need adjusting depending on how balanced this objective is in practice.